Removed need for multiple Azure class helpers
Updated the RubyDoc on Classes and Methods

Fixes #9
